requirements.txt
gevent
gevent-websocket
paramiko
arrow
beaker
psycopg2-binary
pandas
numpy
pymssql
orjson
urllib3
fastapi
schedule
whitenoise
ldap3
wsaccel
pysftp
xmltodict
ujson
genson
azure-storage-blob
azure-identity
progress
dict2xml
requests
certifi
googlemaps
msal
cryptography
docker
pygeocoder
google-cloud-bigquery
pandas-gbq
matplotlib
ndjson
pyodbc
office365-rest-python-client
office365
openpyxl
xlrd
tabula-py
